+++ docker_cli/cli/__init__.py
@@ -1,3 +1,1 @@
 from .cli import main
-
-main()
